requirements.txt
torch
numpy
pandas
matplotlib
regex
streamlit
scikit-learn
tokenizer